requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-extensions==2.1.6
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.2
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mysqlclient==1.4.2
numpy==1.16.2
odfpy==1.4.0
openpyxl==2.6.1
pandas==0.24.2
psutil==5.6.1
pydot==1.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
scipy==1.2.1
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
